+++ src/Stratis.Bitcoin.Features.PoA/FederationHistory.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using NBitcoin;
 using NBitcoin.Crypto;
 using Stratis.Bitcoin.Features.PoA.Voting;
+using Stratis.Bitcoin.Utilities;
 
 namespace Stratis.Bitcoin.Features.PoA
 {
@@ -17,21 +18,6 @@ namespace Stratis.Bitcoin.Features.PoA
         /// <returns>The federation member or <c>null</c> if the member could not be determined.</returns>
         /// <exception cref="ConsensusErrorException">In case timestamp is invalid.</exception>
         IFederationMember GetFederationMemberForBlock(ChainedHeader chainedHeader);
-
-        /// <summary>Gets the miner for a specified block by first looking at a cache 
-        /// and then the signature in <see cref="PoABlockHeader.BlockSignature"/>.</summary>
-        /// <param name="chainedHeader">Identifies the block and timestamp.</param>
-        /// <param name="federation">The federation members at the block height.</param>
-        /// <returns>The federation member or <c>null</c> if the member could not be determined.</returns>
-        IFederationMember GetFederationMemberForBlock(ChainedHeader chainedHeader, List<IFederationMember> federation);
-
-        IFederationMember GetFederationMemberForBlock(PoABlockHeader blockHeader, List<IFederationMember> federation, bool votingManagerV2);
-
-        /// <summary>Gets the miner (and federation) for the specified blocks by first looking at a cache 
-        /// and then the signature in <see cref="PoABlockHeader.BlockSignature"/>.</summary>
-        /// <param name="chainedHeaders">Identifies the blocks and timestamps.</param>
-        /// <returns>The federation member or <c>null</c> if the member could not be determined, as well as the federations.</returns>
-        (IFederationMember[] miners, (List<IFederationMember> members, HashSet<IFederationMember> whoJoined)[] federations) GetFederationMembersForBlocks(ChainedHeader[] chainedHeaders);
 
         /// <summary>Gets the federation for a specified block.</summary>
         /// <param name="chainedHeader">Identifies the block and timestamp.</param>
@@ -45,10 +31,19 @@ namespace Stratis.Bitcoin.Features.PoA
         IFederationMember GetFederationMemberForTimestamp(uint headerUnixTimestamp, PoAConsensusOptions poAConsensusOptions, List<IFederationMember> federationMembers = null);
 
         /// <summary>
-        /// Determines the height from which the voting manager v2 is active.
+        /// Determines when a federation member was last active. This includes mining or joining.
         /// </summary>
-        /// <returns>The height from which the voting manager v2 is active.</returns>
-        int GetVotingManagerV2ActivationHeight();
+        /// <param name="federationMember">Member to check activity of.</param>
+        /// <param name="blockHeader">Block at which to check for past activity.</param>
+        /// <param name="lastActiveTime">Time at which member was last active.</param>
+        /// <returns><c>True</c> if the information is available and <c>false</c> otherwise.</returns>
+        bool GetLastActiveTime(IFederationMember federationMember, ChainedHeader blockHeader, out uint lastActiveTime);
+
+        /// <summary>
+        /// Determines when the current federation members were last active. This includes mining or joining.
+        /// </summary>
+        /// <returns>A list of public keys and the times at which they were active.</returns>
+        ConcurrentDictionary<IFederationMember, uint> GetFederationMembersByLastActiveTime();
     }
 
     /// <summary>
@@ -60,30 +55,62 @@ namespace Stratis.Bitcoin.Features.PoA
     {
         private readonly IFederationManager federationManager;
         private readonly VotingManager votingManager;
+        private readonly ChainIndexer chainIndexer;
         private readonly Network network;
         private readonly ConcurrentDictionary<uint256, PubKey[]> candidatesByBlockHash;
+        private readonly object lockObject;
 
-        public FederationHistory(IFederationManager federationManager, Network network, VotingManager votingManager = null)
+        private SortedDictionary<uint, (List<IFederationMember>, IFederationMember)> federationHistory;
+        private ConcurrentDictionary<PubKey, List<uint>> lastActiveTimeByPubKey;
+        private ChainedHeader lastActiveTip;
+
+        public FederationHistory(IFederationManager federationManager, Network network, VotingManager votingManager = null, ChainIndexer chainIndexer = null)
         {
             this.federationManager = federationManager;
             this.votingManager = votingManager;
+            this.chainIndexer = chainIndexer;
             this.network = network;
             this.candidatesByBlockHash = new ConcurrentDictionary<uint256, PubKey[]>();
+            this.lockObject = new object();
+            this.lastActiveTimeByPubKey = new ConcurrentDictionary<PubKey, List<uint>>();
+            this.federationHistory = new SortedDictionary<uint, (List<IFederationMember>, IFederationMember)>();
+            this.lastActiveTip = null;
         }
 
         /// <inheritdoc />
         public List<IFederationMember> GetFederationForBlock(ChainedHeader chainedHeader)
         {
-            return (this.votingManager == null) ? this.federationManager.GetFederationMembers() : this.votingManager.GetModifiedFederation(chainedHeader);
+            lock (this.lockObject)
+            {
+                this.UpdateTip(chainedHeader);
+
+                if (this.federationHistory.TryGetValue(chainedHeader.Header.Time, out (List<IFederationMember> modifiedFederation, IFederationMember miner) item))
+                    return item.modifiedFederation;
+
+                return (this.votingManager == null) ? this.federationManager.GetFederationMembers() : this.votingManager.GetModifiedFederation(chainedHeader);
+            }
         }
 
         /// <inheritdoc />
         public IFederationMember GetFederationMemberForBlock(ChainedHeader chainedHeader)
         {
-            return GetFederationMemberForBlock(chainedHeader, this.GetFederationForBlock(chainedHeader));
+            lock (this.lockObject)
+            {
+                this.UpdateTip(chainedHeader);
+
+                if (this.federationHistory.TryGetValue(chainedHeader.Header.Time, out (List<IFederationMember> modifiedFederation, IFederationMember miner) item))
+                    return item.miner;
+
+                List<IFederationMember> federation = (this.votingManager == null) ? this.federationManager.GetFederationMembers() : this.votingManager.GetModifiedFederation(chainedHeader);
+
+                if (chainedHeader.Height == 0)
+                    return federation.Last();
+
+                return GetFederationMemberForBlock(chainedHeader.Header as PoABlockHeader, federation, chainedHeader.Height >= this.GetVotingManagerV2ActivationHeight());
+            }
         }
 
-        public int GetVotingManagerV2ActivationHeight()
+        private int GetVotingManagerV2ActivationHeight()
         {
             if (this.network.Consensus.Options is PoAConsensusOptions poaConsensusOptions)
                 return (poaConsensusOptions.VotingManagerV2ActivationHeight == 0) ? int.MaxValue : poaConsensusOptions.VotingManagerV2ActivationHeight;
@@ -91,17 +118,7 @@ namespace Stratis.Bitcoin.Features.PoA
             return 0;
         }
 
-        /// <inheritdoc />
-        public IFederationMember GetFederationMemberForBlock(ChainedHeader chainedHeader, List<IFederationMember> federation)
-        {
-            if (chainedHeader.Height == 0)
-                return federation.Last();
-
-            return GetFederationMemberForBlock(chainedHeader.Header as PoABlockHeader, federation, chainedHeader.Height >= this.GetVotingManagerV2ActivationHeight());
-        }
-
-        /// <inheritdoc />
-        public (IFederationMember[] miners, (List<IFederationMember> members, HashSet<IFederationMember> whoJoined)[] federations) GetFederationMembersForBlocks(ChainedHeader[] chainedHeaders)
+        private (IFederationMember[] miners, (List<IFederationMember> members, HashSet<IFederationMember> whoJoined)[] federations) GetFederationMembersForBlocks(ChainedHeader[] chainedHeaders)
         {
             (List<IFederationMember> members, HashSet<IFederationMember> whoJoined)[] federations = this.votingManager.GetModifiedFederations(chainedHeaders).ToArray();
             IFederationMember[] miners = new IFederationMember[chainedHeaders.Length];
@@ -118,7 +135,7 @@ namespace Stratis.Bitcoin.Features.PoA
             return (miners, federations);
         }
 
-        public IFederationMember GetFederationMemberForBlock(PoABlockHeader blockHeader, List<IFederationMember> federation, bool votingManagerV2)
+        private IFederationMember GetFederationMemberForBlock(PoABlockHeader blockHeader, List<IFederationMember> federation, bool votingManagerV2)
         {
             if (!votingManagerV2)
                 return GetFederationMemberForTimestamp(blockHeader.Time, this.network.Consensus.Options as PoAConsensusOptions, federation);
@@ -172,6 +189,152 @@ namespace Stratis.Bitcoin.Features.PoA
         {
             uint roundLength = (uint)(federationMembersCount * poAConsensusOptions.TargetSpacingSeconds);
             return roundLength;
+        }
+
+        /// <inheritdoc />
+        public ConcurrentDictionary<IFederationMember, uint> GetFederationMembersByLastActiveTime()
+        {
+            lock (this.lockObject)
+            {
+                ChainedHeader tip = this.lastActiveTip ?? this.chainIndexer.GetHeader(0);
+
+                List<IFederationMember> federationMembers = this.GetFederationForBlock(tip);
+
+                return new ConcurrentDictionary<IFederationMember, uint>(federationMembers
+                    .Select(m => (m, (this.GetLastActiveTime(m, tip, out uint lastActiveTime) && lastActiveTime != default) ? lastActiveTime : this.network.GenesisTime))
+                    .ToDictionary(x => x.m, x => x.Item2));
+            }
+        }
+
+        private void DiscardActivityAboveTime(uint discardAboveTime)
+        {
+            var remove = new List<PubKey>();
+
+            foreach ((PubKey pubKey, List<uint> activity) in this.lastActiveTimeByPubKey)
+            {
+                int pos = BinarySearch.BinaryFindFirst(x => activity[x] > discardAboveTime, 0, activity.Count);
+                if (pos >= 0)
+                {
+                    if (pos == 0)
+                        remove.Add(pubKey);
+                    else
+                        activity.RemoveRange(pos, activity.Count - pos);
+                }
+            }
+
+            foreach (PubKey pubKey in remove)
+                this.lastActiveTimeByPubKey.Remove(pubKey, out _);
+
+            int pos2 = BinarySearch.BinaryFindFirst(x => this.federationHistory.ElementAt(x).Key > discardAboveTime, 0, this.federationHistory.Values.Count);
+            if (pos2 > 0)
+                this.federationHistory = new SortedDictionary<uint, (List<IFederationMember>, IFederationMember)>(this.federationHistory.Skip(pos2).ToDictionary(x => x.Key, x => x.Value));
+        }
+
+        private void UpdateTip(ChainedHeader blockHeader)
+        {
+            // TODO: Add cleanup of older blocks.
+
+            if (this.lastActiveTip != null)
+            {
+                if (blockHeader == this.lastActiveTip)
+                    return;
+
+                ChainedHeader fork = this.lastActiveTip.FindFork(blockHeader);
+
+                // If the current chain includes the block then do nothing.
+                if (fork == blockHeader)
+                    return;
+
+                // If the fork shows blocks that are not in common then discard those blocks.
+                if (fork != this.lastActiveTip)
+                {
+                    DiscardActivityAboveTime(fork.Header.Time);
+                    this.lastActiveTip = fork;
+                }
+            }
+
+            var federationMemberActivationTime = ((PoAConsensusOptions)this.network.Consensus.Options).FederationMemberActivationTime ?? 0;
+
+            ChainedHeader[] headers = blockHeader
+                .EnumerateToGenesis()
+                .TakeWhile(h => h.HashBlock != this.lastActiveTip?.HashBlock && h.Header.Time >= federationMemberActivationTime)
+                .Reverse().ToArray();
+
+            if (headers.Length != 0)
+            {
+                (IFederationMember[] miners, (List<IFederationMember> members, HashSet<IFederationMember> whoJoined)[] federations) = this.GetFederationMembersForBlocks(headers);
+
+                for (int i = 0; i < headers.Length; i++)
+                {
+                    ChainedHeader header = headers[i];
+
+                    uint headerTime = header.Header.Time;
+
+                    this.federationHistory[headerTime] = (federations[i].members, miners[i]);
+
+                    if (miners[i] != null)
+                    {
+                        if (!this.lastActiveTimeByPubKey.TryGetValue(miners[i].PubKey, out List<uint> minerActivity))
+                        {
+                            minerActivity = new List<uint>();
+                            this.lastActiveTimeByPubKey[miners[i].PubKey] = minerActivity;
+                        }
+
+                        if (minerActivity.LastOrDefault() != headerTime)
+                            minerActivity.Add(headerTime);
+                    }
+
+                    foreach (IFederationMember member in federations[i].whoJoined)
+                    {
+                        if (!this.lastActiveTimeByPubKey.TryGetValue(member.PubKey, out List<uint> joinActivity))
+                        {
+                            joinActivity = new List<uint>();
+                            this.lastActiveTimeByPubKey[member.PubKey] = joinActivity;
+                        }
+
+                        if (joinActivity.LastOrDefault() != headerTime)
+                            joinActivity.Add(headerTime);
+                    }
+                }
+            }
+
+            this.lastActiveTip = blockHeader;
+        }
+
+        /// <inheritdoc />
+        public bool GetLastActiveTime(IFederationMember federationMember, ChainedHeader blockHeader, out uint lastActiveTime)
+        {
+            lock (this.lockObject)
+            {
+                UpdateTip(blockHeader);
+
+                if (this.lastActiveTimeByPubKey.TryGetValue(federationMember.PubKey, out List<uint> activity))
+                {
+                    uint blockTime = blockHeader.Header.Time;
+
+                    lastActiveTime = activity.Last();
+                    if (activity.Last() <= blockTime)
+                        return true;
+
+                    int pos = BinarySearch.BinaryFindFirst(i => activity[i] > blockTime, 0, activity.Count);
+
+                    if (pos > 0)
+                    {
+                        lastActiveTime = activity[pos - 1];
+                        return true;
+                    }
+                }
+
+                if (((PoAConsensusOptions)this.network.Consensus.Options).GenesisFederationMembers.Any(m => m.PubKey == federationMember.PubKey))
+                {
+                    lastActiveTime = default;
+                    return true;
+                }
+
+                // This should never happen.
+                lastActiveTime = default;
+                return false;
+            }
         }
     }
 }
