+++ src/Stratis.Bitcoin.Features.PoA/Voting/FederationController.cs
@@ -18,6 +18,7 @@ namespace Stratis.Bitcoin.Features.PoA.Voting
         private readonly ChainIndexer chainIndexer;
         private readonly IFederationManager federationManager;
         private readonly IIdleFederationMembersKicker idleFederationMembersKicker;
+        private readonly IFederationHistory federationHistory;
         private readonly ILogger logger;
         private readonly Network network;
         private readonly ReconstructFederationService reconstructFederationService;
@@ -29,11 +30,13 @@ namespace Stratis.Bitcoin.Features.PoA.Voting
             VotingManager votingManager,
             Network network,
             IIdleFederationMembersKicker idleFederationMembersKicker,
+            IFederationHistory federationHistory,
             ReconstructFederationService reconstructFederationService)
         {
             this.chainIndexer = chainIndexer;
             this.federationManager = federationManager;
             this.idleFederationMembersKicker = idleFederationMembersKicker;
+            this.federationHistory = federationHistory;
             this.network = network;
             this.reconstructFederationService = reconstructFederationService;
             this.votingManager = votingManager;
@@ -86,7 +89,7 @@ namespace Stratis.Bitcoin.Features.PoA.Voting
                     PubKey = this.federationManager.CurrentFederationKey.PubKey.ToHex()
                 };
 
-                KeyValuePair<IFederationMember, uint> lastActive = this.idleFederationMembersKicker.GetFederationMembersByLastActiveTime().FirstOrDefault(x => x.Key.PubKey == this.federationManager.CurrentFederationKey.PubKey);
+                KeyValuePair<IFederationMember, uint> lastActive = this.federationHistory.GetFederationMembersByLastActiveTime().FirstOrDefault(x => x.Key.PubKey == this.federationManager.CurrentFederationKey.PubKey);
                 if (lastActive.Key != null)
                 {
                     federationMemberModel.LastActiveTime = new DateTime(1970, 1, 1, 0, 0, 0).AddSeconds(lastActive.Value);
@@ -154,7 +157,7 @@ namespace Stratis.Bitcoin.Features.PoA.Voting
                 var federationMemberModels = new List<FederationMemberModel>();
 
                 // Get their last active times.
-                ConcurrentDictionary<IFederationMember, uint> activeTimes = this.idleFederationMembersKicker.GetFederationMembersByLastActiveTime();
+                ConcurrentDictionary<IFederationMember, uint> activeTimes = this.federationHistory.GetFederationMembersByLastActiveTime();
                 foreach (IFederationMember federationMember in federationMembers)
                 {
                     federationMemberModels.Add(new FederationMemberModel()
